Extract AST and Props from flow to separate file.
