+++ compiler/crates/relay-typegen/src/flow.rs
@@ -5,43 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-use interner::{Intern, StringKey};
-use lazy_static::lazy_static;
+use crate::writer::{AST, Prop, SPREAD_KEY};
+use interner::StringKey;
 use std::fmt::{Result, Write};
-
-#[derive(Debug, Clone)]
-pub enum AST {
-    Union(Vec<AST>),
-    Intersection(Vec<AST>),
-    ReadOnlyArray(Box<AST>),
-    Nullable(Box<AST>),
-    Identifier(StringKey),
-    /// Printed as is, should be valid Flow code.
-    RawType(StringKey),
-    String,
-    StringLiteral(StringKey),
-    /// Prints as `"%other" with a comment explaining open enums.
-    OtherEnumValue,
-    Local3DPayload(StringKey, Box<AST>),
-    ExactObject(Vec<Prop>),
-    InexactObject(Vec<Prop>),
-    Number,
-    Boolean,
-    Any,
-}
-
-lazy_static! {
-    /// Special key for `Prop` that turns into an object spread: ...value
-    pub static ref SPREAD_KEY: StringKey = "\0SPREAD".intern();
-}
-
-#[derive(Debug, Clone)]
-pub struct Prop {
-    pub key: StringKey,
-    pub value: AST,
-    pub read_only: bool,
-    pub optional: bool,
-}
 
 pub fn print_type(ast: &AST) -> String {
     let mut printer = Printer {
